Refactor: Create SchemaWriter for schema writing functions
